+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.4.4] - 2019-10-29 WangXH <kurapica125@outlook.com>
+### Added
+- `IsTemporary` Property is added to System.Web.HttpSession, so you can mark a session be cleared when user closed the browser
+
+### Changed
+- Add support for `IsTemporary` session settings
+
 ## [1.4.3] - 2019-10-29 WangXH <kurapica125@outlook.com>
 ### Changed
 - The System.Data.DataEntityCache can use cache class(extend System.Data.ICache) as its second template parameter, so it can create cache object directly, works like `System.Data.DataCollection[{MyDataContext, NgxLua.Redis }]`